ref: Fix circular dependency checks & setup
