+++ package.json
@@ -114,7 +114,7 @@
     "jest-environment-node": "^27.5.1",
     "jsdom": "^21.1.2",
     "lerna": "7.1.1",
-    "madge": "4.0.2",
+    "madge": "7.0.0",
     "nodemon": "^2.0.16",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.1.1",
