+++ packages/bun/package.json
@@ -76,12 +76,5 @@
   "volta": {
     "extends": "../../package.json"
   },
-  "sideEffects": false,
-  "madge": {
-    "detectiveOptions": {
-      "ts": {
-        "skipTypeImports": true
-      }
-    }
-  }
+  "sideEffects": false
 }
