+++ packages/core/package.json
@@ -69,12 +69,5 @@
   "volta": {
     "extends": "../../package.json"
   },
-  "madge": {
-    "detectiveOptions": {
-      "ts": {
-        "skipTypeImports": true
-      }
-    }
-  },
   "sideEffects": false
 }
