+++ packages/core/src/tracing/index.ts
@@ -18,7 +18,11 @@ export {
   withActiveSpan,
   suppressTracing,
 } from './trace';
-export { getDynamicSamplingContextFromClient, getDynamicSamplingContextFromSpan } from './dynamicSamplingContext';
+export {
+  getDynamicSamplingContextFromClient,
+  getDynamicSamplingContextFromSpan,
+  spanToBaggageHeader,
+} from './dynamicSamplingContext';
 export { setMeasurement, timedEventsToMeasurements } from './measurement';
 export { sampleSpan } from './sampling';
 export { logSpanEnd, logSpanStart } from './logSpans';
