+++ packages/deno/package.json
@@ -63,13 +63,6 @@
     "extends": "../../package.json"
   },
   "sideEffects": false,
-  "madge": {
-    "detectiveOptions": {
-      "ts": {
-        "skipTypeImports": true
-      }
-    }
-  },
   "nx": {
     "targets": {
       "build:transpile": {
