+++ packages/node/src/utils/commonjs.ts
@@ -1,0 +1,4 @@
+/** Detect CommonJS. */
+export function isCjs(): boolean {
+  return typeof require !== 'undefined';
+}
