+++ packages/vercel-edge/package.json
@@ -73,12 +73,5 @@
   "volta": {
     "extends": "../../package.json"
   },
-  "sideEffects": false,
-  "madge": {
-    "detectiveOptions": {
-      "ts": {
-        "skipTypeImports": true
-      }
-    }
-  }
+  "sideEffects": false
 }
